projects/calendar_integration/requirements.txt
# Core dependencies
google-api-python-client==2.107.0
google-auth==2.23.0
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1

# Environment and configuration
python-dotenv==1.0.0
pyyaml==6.0.1

# Testing
pytest==7.4.0
pytest-cov==4.1.0
pytest-mock==3.11.1

# Code quality
flake8==6.1.0
black==23.7.0
isort==5.12.0

# Utilities
requests==2.31.0
cryptography==41.0.3  # For secure token storage

# Documentation
sphinx==7.2.6
sphinx-rtd-theme==1.3.0 